% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{aq_app}
\alias{aq_app}
\title{Launch Aquifer Data Shiny App}
\usage{
aq_app()
}
\description{
This app allows you to load a shapefile and filter aquifer/well data
according to region, explore data, and export cleaned files.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
aq_app()
\dontshow{\}) # examplesIf}
}
