% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetching_data.R
\name{cache_clean}
\alias{cache_clean}
\title{Clean cache}
\usage{
cache_clean(bcmaps_cded = FALSE)
}
\arguments{
\item{bcmaps_cded}{Logical. Whether or not to also remove CDED files cached with
the bcmaps package. These are used by bcaquifertools for acquiring TRIM
data, but may also be cached for use by other workflows.}
}
\description{
Removes data cache
}
\examples{

# cache_clean()
# cache_clean(bcmaps_cded = TRUE)

}
