% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Common arguments and documentation for various functions}
\arguments{
\item{region}{sf simple features object. Shape file of the region of
interest.}

\item{update}{Logical. Force update of the data?}

\item{wells_sub}{sf spatial data frame. Subset of wells data output by
\code{wells_subset()}}

\item{permission}{Logical. Permission to create the cache folder. If \code{FALSE},
user is asked for permission, if \code{TRUE}, permission is implied.}
}
\description{
Common arguments and documentation for various functions
}
\details{
Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)
}
\keyword{internal}
