% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawdown.R
\name{drawdown}
\alias{drawdown}
\title{Create Excel Drawdown template file}
\usage{
drawdown(
  location,
  rate = NA,
  duration = NA,
  overwrite = FALSE,
  file_name = NULL,
  update = FALSE
)
}
\arguments{
\item{location}{Numeric. Either a vector with longitude/latitude, or a well
tag number.}

\item{rate}{Numeric. The pumping rate in m3/day (defaults to \code{NA}, fillable in Excel file).}

\item{duration}{Numeric. The duration of pumping in days (defaults to \code{NA}, fillable in Excel file).}

\item{overwrite}{Logical. Overwrite existing file?}

\item{file_name}{Character. Optional file name for created excel file.}

\item{update}{Logical. Force update of the data?}
}
\value{
Creates excel file
}
\description{
\strong{This function is still in development. Use at your own risk!}
Creates an Excel file with formulas for calculating drawdown. File is
pre-filled with wells within 100km of \code{location} (well or coordinates).
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
drawdown(85199, rate = 343, duration = 180)
drawdown(85199, rate = 343, duration = 180, overwrite = TRUE)
drawdown(22966, rate = 343, duration = 180)
drawdown(22966)

drawdown(c(-123.5593, 48.647), rate = 3.97, duration = 180)
drawdown(c(-123.5593, 48.647))
\dontshow{\}) # examplesIf}
}
