% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixes.R
\name{fix_yield_zero}
\alias{fix_yield_zero}
\title{Fix the yield of a well if equal to zero}
\usage{
fix_yield_zero(wells_sub, fix = TRUE)
}
\arguments{
\item{wells_sub}{Data frame. The subsetted Wells data frame.}

\item{fix}{Logical. Whether to apply the fix.}
}
\description{
The \code{flag_yield_zero} flag identifies wells where the yield was marked as 0
in GWELLS but probably should have been recorded as \code{NA}. Fixing these well
yields means replacing the yield of 0 with \code{NA}.
}
