% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flags}
\alias{flags}
\title{Flags}
\format{
\subsection{\code{flags}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{Flag}{flag name}
\item{Description}{Flag description}
\item{Solution}{Suggested solution to the problem}
}
}
}
\usage{
flags
}
\description{
A glossary of flag terms
}
\keyword{datasets}
