% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar.R
\name{lidar_fetch}
\alias{lidar_fetch}
\title{Fetch LiDAR tiles corresponding to a sf object}
\usage{
lidar_fetch(region, out_dir = NULL, only_new = TRUE, verbose = FALSE)
}
\arguments{
\item{region}{sf object}

\item{out_dir}{Character. Output folder, defaults to working directory}

\item{only_new}{Logical. Only download new, missing LiDAR tiles, default to
TRUE}

\item{verbose}{Logical. Show extra output. Can be useful to check on
progress.}
}
\value{
Data frame describing tiles and location on disk
}
\description{
Fetch LiDAR tiles corresponding to a sf object
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

creek <- sf::st_read("misc/data/Clinton_Creek.shp")
lidar_fetch(creek)
\dontshow{\}) # examplesIf}
}
