% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lithology.R
\name{lith_fix}
\alias{lith_fix}
\title{Fix lithology descriptions}
\usage{
lith_fix(file = "lithology.csv", desc = NULL)
}
\arguments{
\item{file}{Character. Lithology file name stored in cache}

\item{desc}{Character. Text string to convert (overrides \code{file}).}
}
\description{
Clean and categorize lithology descriptions into primary, secondary, tertiary
and final lithology categories.
}
\examples{

lith_fix(desc = "sandy gravel")

# basic spell checks
lith_fix(desc = "saandy gravel")

}
