% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiles}
\alias{tiles}
\title{tiles}
\format{
\subsection{\code{tiles}}{

A data frame with 7,129 rows and 5 columns:
\describe{
\item{map_tile}{Tile name}
\item{geometry}{Spatial data}
\item{utm}{Projection}
\item{tile_name}{Lidar tile name}
\item{url}{Link to Lidar tile}

}
}
}
\usage{
tiles
}
\description{
A spatial data frame of map tiles with corresponding links to Lidar tiles.
}
\details{
The spatial grid of map tiles is obtained from the BC Data Catalogue,
\href{https://catalogue.data.gov.bc.ca/dataset/a61976ac-d8e8-4862-851e-d105227b6525}{BCGS 1:20,000 Grid}

Links to Lidar tile urls are extracted from the list at the
\href{https://governmentofbc.maps.arcgis.com/apps/MapSeries/index.html?appid=d06b37979b0c4709b7fcf2a1ed458e03}{LidarBC Open LiDAR Data Portal}
}
\keyword{datasets}
