% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wells_eg_fixed}
\alias{wells_eg_fixed}
\title{wells_eg_fixed}
\format{
\subsection{\code{wells_eg_fixed}}{

A data frame with 31 rows and 71 columns.
}
}
\usage{
wells_eg_fixed
}
\description{
Example data of fixed wells subsetted and with elevation added
}
\keyword{datasets}
