% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wells_eg_unfixed}
\alias{wells_eg_unfixed}
\title{wells_eg_unfixed}
\format{
\subsection{\code{wells_eg_unfixed}}{

A data frame with 31 rows and 71 columns.
}
}
\usage{
wells_eg_unfixed
}
\description{
Example data of unfixed wells subsetted and with elevation added
}
\keyword{datasets}
