% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{wells_subset}
\alias{wells_subset}
\title{Subset wells to region}
\usage{
wells_subset(region, fix_bottom = TRUE, fix_depth = TRUE, update = FALSE)
}
\arguments{
\item{region}{sf simple features object. Shape file of the region of
interest.}

\item{fix_bottom}{Logical. Whether to add 1m to bottom lithology intervals that
has no thickness (identified by \code{flat_int_bottom}). Default \code{TRUE}.}

\item{fix_depth}{Logical. Whether to fix missing well depths by making them
equal to the depth of the final lithology layer. Default \code{TRUE}.}

\item{update}{Logical. Force update of the data?}
}
\description{
Filter the GWELLS data returning only wells within the provided shapefile.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

library(sf)

# Load a shape file defining the region of interest
creek_sf <- st_read("misc/data/Clinton_Creek.shp")

# Get wells within this region
creek_wells <- wells_subset(creek_sf)
\dontshow{\}) # examplesIf}
}
