% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_read}
\alias{data_read}
\title{Download, Update, and/or load data}
\usage{
data_read(type, update = FALSE, permission = FALSE)
}
\arguments{
\item{type}{Character. Type of data to return, one of \code{wells}, \code{wells_sf}, or
\code{lithology}}

\item{update}{Logical. Force update of the data?}
}
\description{
This function downloads, updates or loads locally stored data. Currently this
function returns \code{wells}, \code{wells_sf}, or \code{lithology} data. Note that these
data are originally from GWELLS, but are cleaned and summarized for use in
the bcaquiferdata package. For example \code{wells_sf} is a spatial version of the
data, and \code{lithology} is a cleaned and standardized version of lithology.
\code{wells} also contains the new standardized \code{lithology} data, along with the
original lithology observations and intermediate classification steps to
simplify error tracing.
}
\details{
Under normal circumstances, users will not need to use this function as it is
used internally by the main workflow functions. However, users may wish to
overview entire datasets.

Bear in mind that the lithology cleaning and
standardizing, while better than the original data, will almost certainly
still have errors!
}
\examples{

\dontrun{
wells <- data_read("wells")
}
}
