% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetching_data.R
\name{data_update}
\alias{data_update}
\title{Update cached data}
\usage{
data_update(
  type = c("wells", "lithology"),
  download = TRUE,
  permission = FALSE
)
}
\arguments{
\item{type}{Character. Type of data to update. One of "all", "wells",
"lithology"}

\item{download}{Logical. Whether to re-download and process the data
(\code{TRUE}), or just re-process it (\code{FALSE}).}

\item{permission}{Logical. Permission to create the cache folder. If \code{FALSE},
user is asked for permission, if \code{TRUE}, permission is implied.}
}
\description{
Update the GWELLs data stored locally.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
data_update(type = "wells")
data_update(type = "lithology")
\dontshow{\}) # examplesIf}
}
