% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_update}
\alias{data_update}
\title{Update cached data}
\usage{
data_update(type = "all", download = TRUE, permission = FALSE)
}
\arguments{
\item{type}{Character. Type of data to update. One of "all", "wells",
"lithology"}

\item{download}{Logical. Whether to re-download and process the data
(\code{TRUE}), or just re-process it (\code{FALSE}).}

\item{permission}{Logical. Permission to create the cache folder. If \code{FALSE},
user is asked for permission, if \code{TRUE}, permission is implied.}
}
\description{
Update the GWELLs data stored locally.
}
