% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{dem_region}
\alias{dem_region}
\title{Fetch and trim DEM of a region}
\usage{
dem_region(
  region,
  source = "lidar",
  buffer = 1,
  lidar_dir = NULL,
  only_new = TRUE,
  progress = httr::progress(),
  type
)
}
\arguments{
\item{region}{sf simple features object. Shape file of the region of
interest.}

\item{source}{Character. Source of DEM, "lidar", "trim" or a file path (or
vector of file paths) to a custom DEM file. See Details.}

\item{buffer}{Numeric. Percent buffer to apply to the \code{region} spatial file
before cropping the DEM data to match. Increase this value if you find
that wells on the edge of your area aren't been matched to elevations when
using \code{wells_elev()}.}

\item{lidar_dir}{Character. File path of where Lidar tiles should be stored.
Defaults to the cache directory. Only applies when \code{source = "lidar"}.}

\item{only_new}{Logical. Whether to download all Lidar tiles, or only new
tiles that don't exist locally. Defaults to TRUE. Only apples when \code{source = "lidar"}.}

\item{progress}{Function. Progress bar to use. Generally leave as is.}
}
\value{
stars spatiotemporal array object
}
\description{
This function takes a shape file of a region and creates a DEM of the region.
Lidar data is stored locally as tiles. Tiles are only downloaded if they
don't already exist unless \code{only_new = FALSE}. TRIM data is obtained via the
\code{bcmaps} package and stored locally as tiles. \strong{Note:} TRIM elevation is
coarser than Lidar Use Lidar unless it is missing for your region of
interest.
}
\details{
Lidar tiles are the newest tile available. If you have reason to need a
historical file, contact the team to discuss your use case.
}
\section{Data Source}{


Lidar data is obtained from the LidarBC portal. The \code{tiles} data frame
contains is an internally created data frame listing tiles and their
respective download locations. Tiles to download are selected based on
overlap between map tiles and the provided shapefile (\code{region}). These Lidar
tiles can be browsed and downloaded manually via the
\href{https://governmentofbc.maps.arcgis.com/apps/MapSeries/index.html?appid=d06b37979b0c4709b7fcf2a1ed458e03}{LidarBC Open LiDAR Data Portal}

The grid of map tiles is obtained from the BC Data Catalogue,
\href{https://catalogue.data.gov.bc.ca/dataset/a61976ac-d8e8-4862-851e-d105227b6525}{BCGS 1:20,000 Grid}

TRIM data is obtained via the \code{bcmaps} package from the BC government
\href{https://catalogue.data.gov.bc.ca/dataset/7b4fef7e-7cae-4379-97b8-62b03e9ac83d}{Data Catalogue}
based on overlap between map tiles and the provided shapefile (\code{region}).

If a file path or vector of file paths are provided, a local DEM is
loaded with \code{stars} and combined with \code{stars::st_mosaic()}. Note that it is
assumed the data is elevation in metres.
}

\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(sf)

# Load a shape file defining the region of interest
creek_sf <- st_read("misc/data/Clinton_Creek.shp")

# Fetch Lidar DEM
creek_lidar <- dem_region(creek_sf)

plot(creek_lidar)

# Fetch TRIM DEM
creek_trim <- dem_region(creek_sf, source = "trim")

plot(creek_trim)

# Use local DEM
koksilah_sf <- st_read("misc/data/Koksilah_watershed4/Koksilah_watershed4.shp")
koksilah_dem <- dem_region(
  koksilah_sf, source = "misc/data/Koksilah_Watershed_DEM_2km_Buffer.tif")
\dontshow{\}) # examplesIf}
}
