% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{fix_depth_mismatch}
\alias{fix_depth_mismatch}
\title{Fix the depth of the well if not equal to final lithology}
\usage{
fix_depth_mismatch(wells_sub)
}
\arguments{
\item{wells_sub}{sf spatial data frame. Subset of wells data output by
\code{wells_subset()}}
}
\value{
Fixed wells_sub data frame
}
\description{
The \code{flag_depth_mismatch} flag identifies wells where the depth is not the
same as the depth of the final lithology layer. Fixing these well depths
means replacing the well depth with the depth of the final lithology layer.
}
\details{
This is done automatically for leapfrog exports, but otherwise not.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}

library(sf)

# Load a shape file defining the region of interest
creek_sf <- st_read("misc/data/Clinton_Creek.shp")

# Get wells within this region
creek_wells <- wells_subset(creek_sf)

# Fix well depths
creek_wells_fixed <- fix_depth_mismatch(creek_wells)

# Explore all fixes
dplyr::select(creek_wells_fixed, dplyr::starts_with("fix"))
\dontshow{\}) # examplesIf}
}
