% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lithology.R
\name{lith_fix}
\alias{lith_fix}
\title{Fix lithology descriptions}
\usage{
lith_fix(desc = NULL)
}
\arguments{
\item{desc}{Character. Text string to convert/fix.}
}
\value{
Data frame of lithology categorizations
}
\description{
Clean and categorize lithology descriptions into primary, secondary, tertiary
and final lithology categories. Generally this function is used internally
when loading and cleaning GWELLS lithology.
}
\details{
However statements can be tested directly with this function to see how it
works and for troubleshooting.
}
\examples{

lith_fix("sandy gravel")

# basic spell checks
lith_fix("saandy gravel")

}
