% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{tiles}
\alias{tiles}
\title{Tiles}
\format{
An object of class \code{sf} (inherits from \code{bcdc_sf}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 7129 rows and 5 columns.
}
\usage{
tiles
}
\description{
A dataset of Lidar tiles available
}
\keyword{datasets}
