% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{wells_elev}
\alias{wells_elev}
\title{Subset wells and add elevation}
\usage{
wells_elev(region, lidar)
}
\arguments{
\item{region}{sf simple features object. Shape file of the region of
interest.}

\item{lidar}{stars simple features object. Output of \code{lidar_region()}.}
}
\value{

}
\description{
This function takes a region shape file and LiDAR object of a region (output
of \code{lidar_region()}), subsets the wells data (from GWELLS) to this region and
adds LiDAR elevation data.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(sf)

# Load a shape file defining the region of interest
creek <- st_read("misc/data/Clinton_Creek.shp")

# Fetch LiDAR DEM
creek_lidar <- lidar_region(creek)

# Collect wells in this region with added elevation from LiDAR
creek_wells <- wells_elev(creek, creek_lidar)

library(ggplot2)
ggplot() +
  geom_sf(data = creek) +
  geom_sf(data = creek_wells, size=0.5, colour = "dark blue",
          fill="NA",show.legend=FALSE) +
 coord_sf(datum = st_crs(3005)) # BC Albers
\dontshow{\}) # examplesIf}
}
