% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{wells_export}
\alias{wells_export}
\title{Export wells data for use in Strater and Voxler}
\usage{
wells_export(wells_sub, id, dir = ".")
}
\arguments{
\item{wells_sub}{Data frame. Output of \code{wells_elev()}}

\item{id}{Character. Id to prepend to all output files e.g., "id_lith.csv"}

\item{dir}{Character. Directory where files should be exported to. Defaults
to working directory.}
}
\description{
Export wells data for use in Strater and Voxler
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

library(sf)

# Load a shape file defining the region of interest
creek <- st_read("misc/data/Clinton_Creek.shp")

# Fetch LiDAR DEM
creek_lidar <- lidar_region(creek)

# Collect wells in this region with added elevation from LiDAR
creek_wells <- wells_elev(creek, creek_lidar)

# Export data for Strater and Voxler
wells_export(creek_wells, id = "clinton")
\dontshow{\}) # examplesIf}
}
