% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{wells_yield}
\alias{wells_yield}
\title{Add yield lithology data to wells subset}
\usage{
wells_yield(wells_sub)
}
\arguments{
\item{wells_sub}{}
}
\description{
Add yield lithology data to wells subset
}
